Made argument defaults default to 'blank' value
Thought about this for awhile and I think that having a jinja2 template
spit out 'None' as a string for arguments with no default is
non-intuitive. So I swapped all default values over to falsy versions of
themselves.

As an example, strings without default arguments will now default to ''
instead of 'None'.